+++ MemoryGame-1B/Src/Managers/GraphqlManager.cs
@@ -1,8 +1,7 @@
 ﻿using System.Threading.Tasks;
 using GraphQL.Client;
 using GraphQL.Common.Request;
-using MemoryGame_1B.SaveData;
-using MemoryGame_1B.Views;
+using MemoryGame_1B.Models;
 
 namespace MemoryGame_1B.Managers
 {
@@ -14,7 +13,8 @@ namespace MemoryGame_1B.Managers
         /// <summary>
         /// The graphql client
         /// </summary>
-        public static readonly GraphQLClient GraphQlClient = new GraphQLClient("https://sleepy-falls-91203.herokuapp.com/graphql");
+        public static readonly GraphQLClient GraphQlClient =
+            new GraphQLClient("https://sleepy-falls-91203.herokuapp.com/graphql");
 
         /// <summary>
         /// Gets the servers
@@ -155,142 +155,6 @@ namespace MemoryGame_1B.Managers
 
             var graphQlResponse = await GraphQlClient.PostAsync(graphQlRequest);
             return graphQlResponse.GetDataFieldAs<CreateCardDataResponse>("createDataGrid");
-        }
-    }
-
-    /// <summary>
-    /// GetServersResponse object
-    /// </summary>
-    public class GetServersResponse
-    {
-        /// <summary>
-        /// The servers
-        /// </summary>
-        public Server[] Servers { get; set; }
-
-        /// <summary>
-        /// The response status
-        /// </summary>
-        public ResponseStatus ResponseStatus { get; set; }
-
-        /// <summary>
-        /// Deconstructs this object
-        /// </summary>
-        /// <param name="servers"></param>
-        /// <param name="responseStatus"></param>
-        public void Deconstruct(out Server[] servers, out ResponseStatus responseStatus)
-        {
-            servers = Servers;
-            responseStatus = ResponseStatus;
-        }
-    }
-
-    /// <summary>
-    /// JoinServerResponse object
-    /// </summary>
-    public class JoinServerResponse
-    {
-        /// <summary>
-        /// The card data for that game
-        /// </summary>
-        public string CardData { get; set; }
-
-        /// <summary>
-        /// The response status
-        /// </summary>
-        public ResponseStatus ResponseStatus { get; set; }
-
-        /// <summary>
-        /// Deconstructs this object
-        /// </summary>
-        /// <param name="cardData"></param>
-        /// <param name="responseStatus"></param>
-        public void Deconstruct(out string cardData, out ResponseStatus responseStatus)
-        {
-            cardData = CardData;
-            responseStatus = ResponseStatus;
-        }
-    }
-
-    /// <summary>
-    /// CreateCardDataResponse object
-    /// </summary>
-    public class CreateCardDataResponse
-    {
-        /// <summary>
-        /// The response status
-        /// </summary>
-        public ResponseStatus ResponseStatus { get; set; }
-
-        /// <summary>
-        /// Deconstructs this object
-        /// </summary>
-        /// <param name="responseStatus"></param>
-        public void Deconstruct(out ResponseStatus responseStatus)
-        {
-            responseStatus = ResponseStatus;
-        }
-    }
-
-    /// <summary>
-    /// DeleteServerResponse object
-    /// </summary>
-    public class DeleteServerResponse
-    {
-        /// <summary>
-        /// The response status
-        /// </summary>
-        public ResponseStatus ResponseStatus { get; set; }
-
-        /// <summary>
-        /// Deconstructs this object
-        /// </summary>
-        /// <param name="responseStatus"></param>
-        public void Deconstruct(out ResponseStatus responseStatus)
-        {
-            responseStatus = ResponseStatus;
-        }
-    }
-
-    /// <summary>
-    /// ResponseStatus of the response
-    /// </summary>
-    public enum ResponseStatus
-    {
-        /// <summary>
-        /// Ok status
-        /// </summary>
-        Ok,
-        /// <summary>
-        /// Error Status
-        /// </summary>
-        Error
-    }
-
-    /// <summary>
-    /// CreateServerResponse of the response
-    /// </summary>
-    public class CreateServerResponse
-    {
-        /// <summary>
-        /// The id
-        /// </summary>
-        public string Id { get; set; }
-
-        /// <summary>
-        /// the response status
-        /// </summary>
-        public ResponseStatus ResponseStatus { get; set; }
-
-        /// <summary>
-        /// Deconstructs this object
-        /// </summary>
-        /// <param name="id"></param>
-        /// <param name="responseStatus"></param>
-        public void Deconstruct(out string id, out ResponseStatus responseStatus)
-        {
-            id = Id;
-            responseStatus = ResponseStatus;
         }
     }
 }
