+++ MemoryGame-1B/Src/Models/ResponseStatus.cs
@@ -1,0 +1,18 @@
+﻿namespace MemoryGame_1B.Models
+{
+    /// <summary>
+    /// ResponseStatus of the response
+    /// </summary>
+    public enum ResponseStatus
+    {
+        /// <summary>
+        /// Ok status
+        /// </summary>
+        Ok,
+
+        /// <summary>
+        /// Error Status
+        /// </summary>
+        Error
+    }
+}
