+++ MemoryGame-1B/Src/SaveData/GridSize.cs
@@ -9,6 +9,7 @@
         /// 4x4
         /// </summary>
         Normal,
+
         /// <summary>
         /// 6x6
         /// </summary>
