+++ MemoryGame-1B/Views/Help.xaml
@@ -79,6 +79,5 @@
                 </Grid>
             </Grid>
         </Grid>
-
     </Grid>
 </Page>
