+++ MemoryGame-1B/Views/HighScores.xaml
@@ -1,16 +1,15 @@
 ﻿<Page x:Class="MemoryGame_1B.Views.HighScores"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
       xmlns:local="clr-namespace:MemoryGame_1B.Views"
-      mc:Ignorable="d" 
+      mc:Ignorable="d"
       d:DesignHeight="450" d:DesignWidth="800"
       Title="HighScores">
-
     <Grid>
         <Grid.Background>
-            <ImageBrush ImageSource="/Images/Ui/Zombie/InGameBackground.png"/>
+            <ImageBrush ImageSource="/Images/Ui/Zombie/InGameBackground.png" />
         </Grid.Background>
         <Grid.ColumnDefinitions>
             <ColumnDefinition Width="10*" />
@@ -47,28 +46,28 @@
                     <RowDefinition Height="10*" />
                 </Grid.RowDefinitions>
                 <Viewbox Grid.Column="1" VerticalAlignment="Center" Margin="0">
-                    <TextBlock MinWidth="130" HorizontalAlignment="Center" FontWeight="Bold" FontSize="20" >Speler</TextBlock>
+                    <TextBlock MinWidth="130" HorizontalAlignment="Center" FontWeight="Bold" FontSize="20">Speler</TextBlock>
                 </Viewbox>
                 <Viewbox Grid.Column="2" Grid.Row="0" Margin="40,0,0,0" VerticalAlignment="Center">
-                        <TextBlock MinWidth="130" HorizontalAlignment="Center" FontWeight="Bold" FontSize="20">Punten</TextBlock>
-                    </Viewbox>
+                    <TextBlock MinWidth="130" HorizontalAlignment="Center" FontWeight="Bold" FontSize="20">Punten</TextBlock>
+                </Viewbox>
                 <Grid Column="1" Row="2">
-                <Grid.RowDefinitions>
-                    <RowDefinition Height="10*" />
-                    <RowDefinition Height="10*" />
-                    <RowDefinition Height="10*" />
-                </Grid.RowDefinitions>
-                <Viewbox Grid.Column="2" Grid.Row="0" VerticalAlignment="Top">
+                    <Grid.RowDefinitions>
+                        <RowDefinition Height="10*" />
+                        <RowDefinition Height="10*" />
+                        <RowDefinition Height="10*" />
+                    </Grid.RowDefinitions>
+                    <Viewbox Grid.Column="2" Grid.Row="0" VerticalAlignment="Top">
                         <TextBlock MinWidth="130" HorizontalAlignment="Center" FontSize="15">Playername</TextBlock>
-                </Viewbox>
-                <Viewbox Grid.Column="2" Grid.Row="1" VerticalAlignment="Top">
+                    </Viewbox>
+                    <Viewbox Grid.Column="2" Grid.Row="1" VerticalAlignment="Top">
                         <TextBlock MinWidth="130" HorizontalAlignment="Center" FontSize="15">Playername</TextBlock>
-                </Viewbox>
-                <Viewbox Grid.Column="2" Grid.Row="2" VerticalAlignment="Top">
+                    </Viewbox>
+                    <Viewbox Grid.Column="2" Grid.Row="2" VerticalAlignment="Top">
                         <TextBlock MinWidth="130" HorizontalAlignment="Center" FontSize="15">Playername</TextBlock>
-                </Viewbox>
-            </Grid>
-            <Grid Column="2" Row="2" Margin="40,0,0,0">
+                    </Viewbox>
+                </Grid>
+                <Grid Column="2" Row="2" Margin="40,0,0,0">
                     <Grid.RowDefinitions>
                         <RowDefinition Height="10*" />
                         <RowDefinition Height="10*" />
